reduce overhead in marshalImmutable's node translation
It's unclear to me why we did the Bytes+Cast dance for each cid.Cid in
the links list. Perhaps to do a copy, but that wouldn't make sense as
dagpb only reads the CIDs when marshaling.

We can use the same Cid value directly, which shaves off a significant
amount of allocation overhead.

	name         old time/op    new time/op    delta
	Roundtrip-8    5.20µs ± 1%    4.81µs ± 0%   -7.53%  (p=0.004 n=6+5)

	name         old alloc/op   new alloc/op   delta
	Roundtrip-8    7.44kB ± 0%    7.14kB ± 0%   -4.09%  (p=0.002 n=6+6)

	name         old allocs/op  new allocs/op  delta
	Roundtrip-8       139 ± 0%       119 ± 0%  -14.39%  (p=0.002 n=6+6)